+++ src/Ext/Bullet/Body.cpp
@@ -40,16 +40,17 @@ void BulletExt::ExtData::InterceptBullet(TechnoClass* pSource, WeaponTypeClass* 
 
 	if (canAffect)
 	{
-		auto const pTechnoTypeExt = TechnoTypeExt::ExtMap.Find(pSource->GetTechnoType());
-		auto const pWeaponOverride = pTechnoTypeExt->Interceptor_WeaponOverride.Get(pTypeExt->Interceptable_WeaponOverride.Get(nullptr));
-		bool detonate = !pTechnoTypeExt->Interceptor_DeleteOnIntercept.Get(pTypeExt->Interceptable_DeleteOnIntercept);
+		const auto pTechnoTypeExt = TechnoTypeExt::ExtMap.Find(pSource->GetTechnoType());
+		const auto pInterceptorType = pTechnoTypeExt->InterceptorType.get();
+		const auto pWeaponOverride = pInterceptorType->WeaponOverride.Get(pTypeExt->Interceptable_WeaponOverride.Get(nullptr));
+		bool detonate = !pInterceptorType->DeleteOnIntercept.Get(pTypeExt->Interceptable_DeleteOnIntercept);
 
 		this->DetonateOnInterception = detonate;
 
 		if (pWeaponOverride)
 		{
-			bool replaceType = pTechnoTypeExt->Interceptor_WeaponReplaceProjectile;
-			bool cumulative = pTechnoTypeExt->Interceptor_WeaponCumulativeDamage;
+			bool replaceType = pInterceptorType->WeaponReplaceProjectile;
+			bool cumulative = pInterceptorType->WeaponCumulativeDamage;
 
 			pThis->WeaponType = pWeaponOverride;
 			pThis->Health = cumulative ? pThis->Health + pWeaponOverride->Damage : pWeaponOverride->Damage;
@@ -72,7 +73,7 @@ void BulletExt::ExtData::InterceptBullet(TechnoClass* pSource, WeaponTypeClass* 
 			}
 		}
 
-		if (isIntercepted && !pTechnoTypeExt->Interceptor_KeepIntact)
+		if (isIntercepted && !pInterceptorType->KeepIntact)
 			this->InterceptedStatus = InterceptedStatus::Intercepted;
 	}
 }